Turn off PostgreSQL 'trust' authentication
When trust authentication [1] is specified, PostgreSQL assumes that
anyone who can connect to the server is authorized to access the
database with whatever database user name they specify (even superuser).

Since this action is intended to be used on CI, this is unlikely a
desired behaviour. First, all credentials are known and must be
specified in order to avoid flakes. Second, most commonly folks around
there want to test that secrets are gathered and passed down to the
database server correctly.

This patch turns off 'trust' authentication for the PostgreSQL server.

[1] https://www.postgresql.org/docs/15/auth-trust.html

Resolves #5