BREAKING CHANGE: enforce password authentication
It turns out that PostgreSQL comes with weird default that allows
passwordless authentication for localhost connections [1]. This
essentially means that 'password' input parameter for this action was
ignored.

The 'setup-postgres' action's primary use case is to be used on CI where
most of the time authentication is desired in order to verify that
passwords are passed correctly from applications under test.

This patch enforces password authentication even for localhost
connections, making sure that passwords are verified and not ignored.
This will break everyone who previously passed wrong password or didn't
pass it at all.

[1] https://www.postgresql.org/docs/15/auth-trust.html

Fixes: #5