Fix macOS 13 support
Due to some dark magic used to prepare the macOS 13 runner, the brew
install command fails to link unrelated kegs, such as `python@3.12`.
Fortunately, since `python@3.12` is a dependency of a dependency, and
that dependency is already installed, we can workaround the issue by
forcing Homebrew to do not try upgrading dependencies unless absolutely
neccessary.

Fixes: #27