Add support for missing action runners
Turns out that Windows Server 2019 does not support locale names
specified in the format defined by the POSIX standard. This patch
converts the POSIX format to the one supported by Windows 2019.

In addition to that, this patch adds all supported action runners to
the continue integration, so we can make sure that this action works
properly on every supported platform.

Reported-by: Irena Rindos
Fixes: #25