BREAKING CHANGE: change default locale to C
The `en_US.UTF-8` locale, which this action previously defaulted to, is
not available on all platforms, most notably on Windows 2019. In order
to support a wider range of platforms, we better choose a locale that is
available everywhere. According to PostgreSQL docs, the C locale is
available on everywhere, which means it seemed to be a good candidate
for the default value.

It occurred to me that since locale is actually used to sort textual
data, the `en_US` locale might not work well for other languages. Which
means choosing C as a default locale is as good as `en_US` just to get
started, but users are advised to choose the locale they wish by
creating a database with a proper locale settings or table columns.

This change unblocks Windows 2019 support, and adds some other missing
operating systems, such as Ubuntu 20.04 and macOS 11, to the testing
matrix on CI.

Reported-by: Irena Rindos
Fixes: #25