Add SSL support
There are two cases when PostgreSQL with SSL enabled may come handy. The
first one is when PostgreSQL driver developers want to test SSL support
in their drivers. The second one is when you come to depend on some
self-signed certificates and you want to test your application
end-to-end, i.e. with certificates being used.

Co-authored-by: Ihor Kalnytskyi <ihor@kalnytskyi.com>
Signed-off-by: Ihor Kalnytskyi <ihor@kalnytskyi.com>
Signed-off-by: chandr-andr (Kiselev Aleksandr) <chandr@chandr.net>