Suppress Homebrew warnings
When `brew` is used by this action, it generates a bunch of warnings
which are picked up by the github actions capture system and are shown
to end users. They are noisy and most of them are expected. In order to
avoid alerting end users, this patch suppresses them by passing 'quiet'
parameter to the `brew` binary, unlinking linked keg and suppressing env
hints.